test/integration: rename package and files name from master to controlplane in test/integration
